chore(common): CHECKOUT-000 move to scripts folder
