+++ scripts/webpack/index.js
@@ -1,4 +1,5 @@
 module.exports = {
     getNextVersion: require('./get-next-version'),
     transformManifest: require('./transform-manifest'),
+    packageLoaderRules: require('./package-loader-rule'),
 };
